LayoutTests/fast/canvas/html-in-canvas/texElementImage2D-12arg-crop-and-scale.html

<!DOCTYPE html><!-- webkit-test-runner [ CanvasDrawElementEnabled=true ] -->
<title>WebGL texElementImage2D — 12-arg crop+scale + zero-area no-op (TB8)</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
  body { margin: 0; }
  #cvs { display: block; }
  /* 32×32 split horizontally: left red, right green. */
  #child {
    width: 32px;
    height: 32px;
    background: linear-gradient(to right, rgb(255, 0, 0) 0 50%, rgb(0, 255, 0) 50% 100%);
  }
</style>
<body>
<canvas id="cvs" width="32" height="32" layoutsubtree>
    <div id="child"></div>
</canvas>
<script>
function nextPaint() {
    return new Promise(r => requestAnimationFrame(() => requestAnimationFrame(r)));
}

promise_test(async () => {
    await nextPaint();
    const canvas = document.getElementById("cvs");
    const child = document.getElementById("child");
    const gl = canvas.getContext("webgl");
    assert_not_equals(gl, null);

    const tex = gl.createTexture();
    gl.bindTexture(gl.TEXTURE_2D, tex);
    // 12-arg overload: crop right half (16,0,16,16) then scale to 32×32 destination.
    gl.texElementImage2D(gl.TEXTURE_2D, 0, gl.RGBA,
        /*sx*/ 16, /*sy*/ 0, /*swidth*/ 16, /*sheight*/ 16,
        /*width*/ 32, /*height*/ 32,
        gl.RGBA, gl.UNSIGNED_BYTE, child);

    // Zero-area no-op: width=100, height=0 must not crash and must not overwrite the
    // texture we just uploaded. Pins tex-element-image-2d-basic.html:83-84 semantics.
    gl.texElementImage2D(gl.TEXTURE_2D, 0, gl.RGBA,
        /*width*/ 100, /*height*/ 0,
        gl.RGBA, gl.UNSIGNED_BYTE, child);

    const fbo = gl.createFramebuffer();
    gl.bindFramebuffer(gl.FRAMEBUFFER, fbo);
    gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.TEXTURE_2D, tex, 0);
    assert_equals(gl.checkFramebufferStatus(gl.FRAMEBUFFER), gl.FRAMEBUFFER_COMPLETE);

    const pixels = new Uint8Array(32 * 32 * 4);
    gl.readPixels(0, 0, 32, 32, gl.RGBA, gl.UNSIGNED_BYTE, pixels);

    // Whole 32×32 should be green (right-half crop upscaled to 32×32).
    const a = (8 * 32 + 8) * 4;
    const b = (24 * 32 + 24) * 4;
    for (const [label, i] of [["top-left interior", a], ["bottom-right interior", b]]) {
        assert_equals(pixels[i], 0, `${label}: red`);
        assert_equals(pixels[i + 1], 255, `${label}: green`);
        assert_equals(pixels[i + 2], 0, `${label}: blue`);
        assert_equals(pixels[i + 3], 255, `${label}: alpha`);
    }
}, "12-arg texElementImage2D crops then scales; zero-area follow-up call is a no-op");
</script>
</body>
